+++ protected/humhub/modules/user/widgets/PeopleFilterPicker.php
@@ -124,12 +124,12 @@ class PeopleFilterPicker extends BasePicker
      */
     public function getSuggestions($keyword = '')
     {
-        return Profile::find()->select([
-            'id' => $this->itemKey,
-            'text' => $this->itemKey,
-        ])
+        return User::find()
+            ->select(['id' => $this->itemKey, 'text' => $this->itemKey])
+            ->visible()
+            ->joinWith('profile')
+            ->andWhere(['LIKE', $this->itemKey, $keyword])
             ->groupBy($this->itemKey)
-            ->where(['LIKE', $this->itemKey, $keyword])
             ->limit(100)
             ->asArray()
             ->all();