+++ CHANGELOG.md
@@ -6,6 +6,7 @@ HumHub Changelog
 - Fix #5776: Hide beta modules on disabled filter
 - Fix #5799: Incorrect selection when multiple custom filters are in place 
 - Fix #5780: Don't hide icon "remove item" on picker
+- Fix #5800: Hide data of disabled users on picker filters on "People" page
 
 1.11.3 (June 27, 2022)
 ----------------------